backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
amqp==5.3.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.3.6
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
cssselect2==0.7.0
dataclasses-json==0.6.7
distro==1.9.0
Django==5.0.9
django-cors-headers==4.6.0
django-filter==23.3
django-guardian==2.4.0
django-ledger==0.5.6.2
django-treebeard==4.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.7.0
drf-spectacular==0.26.5
exceptiongroup==1.2.2
Faker==30.8.2
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
hl7==0.4.5
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
impacket==0.12.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.4.2
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langsmith==0.1.144
ldap3==2.9.1
ldapdomaindump==0.9.4
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.23.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ofxtools==0.9.5
openai==1.12.0
orjson==3.10.11
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydyf==0.11.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyphen==0.17.0
pyserial==3.5
pysmb==1.2.9.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
reportlab==4.0.8
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.10.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.4.4
tenacity==8.5.0
text-unidecode==1.3
tinycss2==1.4.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.27.1
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
weasyprint==60.2
webencodings==0.5.1
websockets==14.0
Werkzeug==3.1.3
yarl==1.17.2
zopfli==0.2.3.post1
